puntuaciones introducidas y version final

+++ app/src/main/java/org/jumpingshooter/juegos/feralva/jumpingshooter/AlmacenPuntuaciones.java
@@ -5,7 +5,7 @@ import java.util.Vector;
  * Created by Packard Bell on 08/03/2015.
  */
 public interface AlmacenPuntuaciones {
-    public void guardarPuntuacion(double puntos,String nombre );
+    public void guardarPuntuacion(int puntos,String nombre );
     public Vector<String> listaPuntuaciones(int Cantdad);
 }
